fixed path issue
when running the plotRTT() function pyrate usually returns the error that it can't find the input file, as it is searching for it in the very root directory '/'. This line of code fixes the issue, so that pyrate takes the whole path of the input file and parses that into the function.